+++ node-addon-api/utils/ring_buffer.h
@@ -7,7 +7,7 @@ class RingBuffer
 {
 public:
    RingBuffer() : _size(0), _arr(nullptr){}
-   RingBuffer(int bufferSize) : _size(bufferSize + 1), _arr(new TValue[bufferSize + 1]){}
+   RingBuffer(size_t bufferSize) : _size(bufferSize + 1), _arr(new TValue[bufferSize + 1]){}
    ~RingBuffer(){ delete[] _arr;}
 
    void push(const TValue &data)
@@ -25,17 +25,17 @@ public:
 
    bool empty(){ return _start == _end;}
    bool full(){ return (_end + 1) % _size == _start;}
-   int size(){ return (_end - _start + _size) % _size;}
-   int max_size(){ return _size - 1;}
+   size_t size(){ return (_end - _start + _size) % _size;}
+   size_t max_size(){ return _size - 1;}
 
-   void resize(int newBufferSize)
+   void resize(size_t newBufferSize)
    {
       const auto newSize = newBufferSize + 1;
       if (newSize < _size)
          return;
       TValue *newArr = new TValue[newSize];
-      int i = 0;
-      for (int j = _start; j != _end; j = (j + 1) % _size)
+      size_t i = 0;
+      for (size_t j = _start; j != _end; j = (j + 1) % _size)
          newArr[i++] = _arr[j];
       delete[] _arr;
       _arr = newArr;
@@ -43,20 +43,19 @@ public:
       _end = i;
       _size = newSize;
    }
-
-   TValue& get(int index){ return _arr[indexToReal(index)];}
-   TValue& operator[](int index){ return _arr[indexToReal(index)];}
+   TValue& get(size_t index){ return _arr[indexToReal(index)];}
+   TValue& operator[](size_t index){ return _arr[indexToReal(index)];}
 
    // 获取第index个元素的真实位置
-   int indexToReal(int index){ return (_start + index) % _size;}
+   size_t indexToReal(size_t index){ return (_start + index) % _size;}
    // 获取真实位置的元素对应的index
-   int realToIndex(int real){ return (real - _start + _size) % _size;}
+   size_t realToIndex(size_t real){ return (real - _start + _size) % _size;}
 
 protected:
    TValue *_arr;
-   int _start = 0;
-   int _end = 0;
-   int _size;
+   size_t _start = 0;
+   size_t _end = 0;
+   size_t _size;
 };
 
 template <typename TValue>
@@ -64,7 +63,7 @@ class SearchableRingBuffer : public RingBuffer<TValue>
 {
 public:
    SearchableRingBuffer() : RingBuffer<TValue>(){}
-   SearchableRingBuffer(int bufferSize) : RingBuffer<TValue>(bufferSize){}
+   SearchableRingBuffer(size_t bufferSize) : RingBuffer<TValue>(bufferSize){}
 
    void push(TValue data)
    {
@@ -89,11 +88,11 @@ public:
       return _set.find(data) != _set.end();
    }
 
-   void resize(int newBufferSize)
+   void resize(size_t newBufferSize)
    {
       RingBuffer<TValue>::resize(newBufferSize);
       _set.clear();
-      for (int i = _start; i != _end; i = (i + 1) % _size)
+      for (size_t i = _start; i != _end; i = (i + 1) % _size)
          _set.insert(_arr[i]);
    }
